refactor(mobile): extract the bottom-sheet shell + sheet rows
ResourcesSheet (576) and TasksSheet (504) were the two largest live files. They
were also near-duplicates above the content: the trigger, count badge, backdrop,
grabber, header, action button and scroll styles were character-for-character
identical in both, so the sheet chrome existed twice.

Lift it into shared/components/BottomSheet — BottomSheet, SheetTrigger,
SheetActionButton, SheetScroll, SheetEmpty — and move the rows to their feature
modules (ResourceRows, ChannelTaskRow). The sheets drop 576->226 and 504->229.

BottomSheet is CONTROLLED rather than owning its open state: TasksSheet closes
itself before navigating to a task, so that state has to stay reachable from
outside.

The one real difference between the two is keyboard handling — TasksSheet holds
a text input and needs the KeyboardAvoidingView, the dropped nav-bar inset while
the keyboard is up, and flexShrink so a tall keyboard cannot push the sheet off
screen. That is a `keyboardAware` opt-in, so ResourcesSheet keeps exactly the
tree it had rather than gaining a KAV it never wanted.

PendingUploadRow's retryable check now calls isRetryableStatus from sync-core
instead of re-spelling `error || awaiting_network`, so the badge and the upload
manager's own retry logic cannot disagree.

ChannelTaskRow is deliberately not MyTaskRow: that one carries a breadcrumb to
locate a task across projects, where here the channel is already the context.

Behavior-preserving, and checked rather than assumed: a script compared every
moved style against both originals at HEAD — all 10 shared chrome styles, both
action buttons, the sheet/flexShrink composition, and all 17 row styles match
byte-for-byte. Typecheck holds at the 85-error baseline with the two remaining
ResourcesSheet errors being the pre-existing casts, 22 tests pass, no new lint.
